Split server-entry.ts into domain-scoped bootstrap modules
server-entry.ts had grown to 1806 lines, owning ~110 /api route
registrations, request-forwarding glue, raw DB queries, CSP/security
header construction, and background-job aggregation for every feature
area in the app. Move each concern into its own module under
src/server/ (federation-middleware, security-headers, background-jobs,
media-routes, content-negotiation-router) and split API routes into
per-domain registrars under src/server/api/ (auth, user, group, event,
discussion, event-notice, poll, place, admin/*, misc), mirroring the
existing controllers/ domain layout. Also extract the three raw
Drizzle queries that lived in the bootstrap file into new minimal
repositories/actors.ts and repositories/polls.ts.

This is a mechanical, behavior-preserving move: route paths, methods,
forwarding logic, and app.use ordering are unchanged. server-entry.ts
is now a ~35-line composition root.
