+++ src/server/federation-middleware.ts
@@ -1,0 +1,7 @@
+import { integrateFederation } from "@fedify/h3";
+import type { EventHandler } from "h3";
+import { federation } from "~/server/fediverse/federation";
+
+export function createFederationMiddleware(): EventHandler {
+  return integrateFederation(federation, () => undefined);
+}
